extract math_funcs expressions to folders based on spark grouping
