+++ native/spark-expr/src/lib.rs
@@ -29,8 +29,6 @@ mod bitwise_not;
 pub use bitwise_not::{bitwise_not, BitwiseNotExpr};
 mod avg_decimal;
 pub use avg_decimal::AvgDecimal;
-mod checkoverflow;
-pub use checkoverflow::CheckOverflow;
 mod correlation;
 pub use correlation::Correlation;
 mod covariance;
@@ -50,9 +48,6 @@ pub use stddev::Stddev;
 mod structs;
 mod sum_decimal;
 pub use sum_decimal::SumDecimal;
-mod negative;
-pub use negative::{create_negate_expr, NegativeExpr};
-mod normalize_nan;
 mod temporal;
 
 pub mod test_common;
@@ -60,8 +55,9 @@ pub mod timezone;
 mod to_json;
 mod unbound;
 pub use unbound::UnboundColumn;
+mod math_funcs;
 pub mod utils;
-pub use normalize_nan::NormalizeNaNAndZero;
+pub use math_funcs::*;
 
 mod variance;
 pub use variance::Variance;
